% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_helpers.R
\name{ao_build_block_objective}
\alias{ao_build_block_objective}
\title{Block objective function}
\usage{
ao_build_block_objective(partition, objective)
}
\arguments{
\item{partition}{(\code{Partition})\cr
The parameter partition for alternating optimization.
Can be created via \code{\link{Partition}}.}

\item{objective}{(\code{Objective})\cr
The objective function to be optimized via alternating optimization.
Can be created via \code{\link[optimizeR]{Objective}}.}
}
\value{
TODO
}
\description{
This helper function builds the block objective function for the alternating
optimization procedure.
}
\keyword{internal}
