% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.R
\name{ao_input_check}
\alias{ao_input_check}
\title{Input checks}
\usage{
ao_input_check(
  argument_name,
  check_result,
  error_message = check_result,
  prefix = "Input {.var {argument_name}} is bad:"
)
}
\arguments{
\item{argument_name}{(\code{character(1)})\cr
The name of the argument that is checked.}

\item{check_result}{(\code{logical(1)})\cr
\code{TRUE} if the check was successful.}

\item{error_message}{(\code{character(1)})\cr
An error message to be printed.}

\item{prefix}{(\code{character(1)})\cr
A prefix before the error message.}
}
\value{
Either throws an error, or invisible \code{TRUE}.
}
\description{
This helper function checks the inputs for the \code{\link{ao}} function.
}
