% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_helpers.R
\name{ao_input_checks}
\alias{ao_input_checks}
\title{Input checks}
\usage{
ao_input_checks(objective, partition, optimizer, initial, procedure)
}
\arguments{
\item{objective}{(\code{Objective})\cr
The objective function to be optimized via alternating optimization.
Can be created via \code{\link[optimizeR]{Objective}}.}

\item{partition}{(\code{Partition})\cr
The parameter partition for alternating optimization.
Can be created via \code{\link{Partition}}.}

\item{optimizer}{(\code{Optimizer})\cr
The optimizer for solving the sub-problems.
Can be created via \code{\link[optimizeR]{Optimizer}}.}

\item{initial}{(\code{numeric()})\cr
The starting parameter values.}

\item{procedure}{(\code{Procedure})\cr
The alternating optimization procedure.
Can be created via \code{\link{Procedure}}.}
}
\value{
Either throws an error, or invisible \code{TRUE}.
}
\description{
This helper function checks the inputs for the \code{\link{ao}} function.
}
\keyword{internal}
