% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric (vector).}

\item{b}{A numeric (vector).}
}
\value{
A numeric.
}
\description{
This function computes the euclidean distance between two numeric vectors.
}
\examples{
euclidean(c(0,0), c(1,1))
}
