% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_random_partition}
\alias{generate_random_partition}
\title{Generated randomized blocks.}
\usage{
generate_random_partition(
  x = seq_len(self$npar),
  p = self$new_block_probability,
  min = self$minimum_block_number
)
}
\arguments{
\item{x}{The parameter indices.}

\item{p}{The probability to generate a new block.}

\item{min}{The minimum number of blocks.}
}
\value{
TODO
}
\description{
Generated randomized blocks.
}
\author{
Siddhartha Chib
}
