% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{Partition}
\alias{Partition}
\title{Partition Object}
\description{
Partition of the target argument for alternating optimization.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{partitions}}{(\code{list})\cr
description}

\item{\code{random}}{(\code{numeric(1)})\cr
description}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Partition-new}{\code{Partition$new()}}
\item \href{#method-Partition-print}{\code{Partition$print()}}
\item \href{#method-Partition-define}{\code{Partition$define()}}
\item \href{#method-Partition-validate}{\code{Partition$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-new"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$new(n, type = "random")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{integer(1)})\cr
The length of the target vector.}

\item{\code{type}}{(\code{character(1)})\cr
\itemize{
\item \code{"random"}
\item \code{"fixed"}
\item \code{"sequential"}
\item \code{"none"}
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-print"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-define"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-define}{}}}
\subsection{Method \code{define()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$define()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$validate()}\if{html}{\out{</div>}}
}

}
}
