% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ao.R
\name{print.ao}
\alias{print.ao}
\title{Print method for \code{ao}}
\usage{
\method{print}{ao}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{ao}}, which is an object of class \code{ao}.}

\item{...}{ignored}
}
\description{
Print method for \code{ao}
}
