% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{rndblocks}
\alias{rndblocks}
\title{rndblocks makes randomized blocks}
\usage{
rndblocks(x = x, p = 0.3)
}
\arguments{
\item{x}{is a sequence starting from 1 to k (the number of parameters)}

\item{p}{is the probability of starting a new block}
}
\description{
rndblocks makes randomized blocks
}
\author{
Siddhartha Chib
}
