% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao-package.R
\docType{package}
\name{ao-package}
\alias{ao-package}
\title{ao: Alternating Optimization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Alternating optimization of (high-dimensional) functions is an iterative procedure for optimizing jointly over all parameters by alternately optimizing parameter subsets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/ao/}
  \item \url{https://github.com/loelschlaeger/ao/}
  \item Report bugs at \url{https://github.com/loelschlaeger/ao/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

}
\keyword{internal}
