% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_helpers.R
\name{ao_build_block_objective}
\alias{ao_build_block_objective}
\title{Block objective function}
\usage{
ao_build_block_objective(partition)
}
\arguments{
\item{partition}{(\code{numeric()})\cr
The definition of the target argument partition for alternating optimization,
can be created via \code{\link{Partition}}.}
}
\value{
TODO
}
\description{
This helper function builds the block objective function for the alternating
optimization procedure.
}
\keyword{internal}
