% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_helpers.R
\name{ao_input_checks}
\alias{ao_input_checks}
\title{Input checks}
\usage{
ao_input_checks(objective, partition, optimizer, initial, procedure)
}
\arguments{
\item{objective}{(\code{numeric()})\cr
The definition of the objective function to be optimized via alternating
optimization, can be created via \code{\link[optimizeR]{Objective}}.}

\item{partition}{(\code{numeric()})\cr
The definition of the target argument partition for alternating optimization,
can be created via \code{\link{Partition}}.}

\item{optimizer}{(\code{Optimizer})\cr
The definition of the base optimizer that solves the optimization problems in
the partitions, can be created via \code{\link[optimizeR]{Optimizer}}.}

\item{initial}{(\code{numeric()})\cr
The starting parameter values for the optimization.}
}
\value{
Either throws an error, or invisible \code{TRUE}.
}
\description{
This helper function checks the inputs for the \code{\link{ao}} function.
}
\keyword{internal}
