% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_random_partition}
\alias{generate_random_partition}
\title{Generated random partition}
\usage{
generate_random_partition(x, p, min)
}
\arguments{
\item{x}{[\code{integer}]\cr
The parameter indices.}

\item{p}{[\code{numeric(1)}]\cr
The probability to generate a new block.}

\item{min}{[\code{integer(1)}]\cr
The minimum number of blocks.}
}
\value{
A \code{list}, a random partition of \code{x}.
}
\description{
This helper function generates a random parameter partition, which is used
for the randomized AO procedure.
}
\author{
Siddhartha Chib
}
