% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_results}
\alias{merge_results}
\title{Merge optimization results}
\usage{
merge_results(
  results,
  minimize = TRUE,
  add_details = TRUE,
  processes = data.frame()
)
}
\arguments{
\item{results}{[\code{list}]\cr
A \code{list} of outputs from \code{\link[ao]{ao}}.}

\item{minimize}{[\code{logical(1)}]\cr
Minimize during the AO process?

If \code{FALSE}, maximization is performed.}

\item{add_details}{[\code{logical(1)}]\cr
Add details about the AO process to the output?}

\item{processes}{[\code{data.frame}]\cr
A \code{data.frame} describing how the different processes were specified.}
}
\value{
A \code{list}, see section "Output value" on the \code{\link[ao]{ao}} page.
}
\description{
This helper function merges the results of multiple AO processes.
}
