% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{Partition}
\alias{Partition}
\title{Partition Object}
\description{
Partition of the target argument for alternating optimization.
}
\examples{
# TODO

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{npar}}{(\code{integer(1)})\cr
The length of the target argument.}

\item{\code{type}}{The partition type, either \code{"random"}, \code{"fixed"}, \code{"sequential"}, or
\code{"none"}. Supports partial matching.}

\item{\code{new_block_probability}}{(\code{numeric(1)})\cr
Only relevant for \code{type = "random"}.
The probability for a new parameter block in random partitions.
Values close to 0 result in larger parameter blocks, values close to 1
result in smaller parameter blocks.}

\item{\code{minimum_block_number}}{(\code{integer(1)})\cr
Only relevant for \code{type = "random"}.
The minimum number of blocks in random partitions.}

\item{\code{block_attributes}}{(\code{list()})\cr
TODO
Return fixed partition.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Partition-new}{\code{Partition$new()}}
\item \href{#method-Partition-print}{\code{Partition$print()}}
\item \href{#method-Partition-get}{\code{Partition$get()}}
\item \href{#method-Partition-define_fixed_partition}{\code{Partition$define_fixed_partition()}}
\item \href{#method-Partition-define_block_attribute}{\code{Partition$define_block_attribute()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-new"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new object of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$new(npar, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{npar}}{(\code{integer(1)})\cr
The length of the target argument.}

\item{\code{type}}{(\code{character(1)})\cr
The partition type, one of:
\itemize{
\item \code{"random"} (random partition in each iteration)
\item \code{"fixed"} (partition defined via \verb{$define_fixed_partition()} method)
\item \code{"sequential"} (each parameter alone)
\item \code{"none"} (no partition, i.e. standard joint optimization)
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-print"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-print}{}}}
\subsection{Method \code{print()}}{
Print details about this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-get"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-get}{}}}
\subsection{Method \code{get()}}{
Return partition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$get(type = self$type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(\code{character(1)})\cr
The partition type, one of:
\itemize{
\item \code{"random"} (random partition in each iteration)
\item \code{"fixed"} (partition defined via \verb{$define_fixed_partition()} method)
\item \code{"sequential"} (each parameter alone)
\item \code{"none"} (no partition, i.e. standard joint optimization)
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-define_fixed_partition"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-define_fixed_partition}{}}}
\subsection{Method \code{define_fixed_partition()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$define_fixed_partition(fixed_partition = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fixed_partition}}{(\code{list()})\cr
Only relevant for \code{type = "fixed"}.
A \code{list} of \code{integer} (vectors), specifying the partition of the
target parameter vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partition-define_block_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-Partition-define_block_attribute}{}}}
\subsection{Method \code{define_block_attribute()}}{
Defines how attributes of the objective function are to be used during
the alternating optimization procedure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partition$define_block_attribute(attribute_name, attribute_definition)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute_name}}{(\code{character(1)})\cr
The name of an attribute of the objective output.}

\item{\code{attribute_definition}}{(\code{function})\cr
The definition how the attribute is to be used. The \code{function} must
have the two arguments \code{x} and \code{y}, where \code{x} is the
attribute value and \code{y} the parameter indices in the current block.}
}
\if{html}{\out{</div>}}
}
}
}
