% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ao.R
\name{print.ao}
\alias{print.ao}
\title{Print method for an object of class \code{ao}.}
\usage{
\method{print}{ao}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{ao}}, which is an object of class \code{ao}.}

\item{...}{Ignored.}
}
\description{
This function is the print method for an object of class \code{ao}.
}
