% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.R
\name{Procedure}
\alias{Procedure}
\title{Procedure Object}
\description{
Details for alternating optimization procedure.
}
\examples{
# TODO

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{verbose}}{(\code{logical(1)})\cr
Whether to print tracing details during the alternating optimization
process.}

\item{\code{minimize}}{(\code{logical(1)})\cr
Whether to minimize during the alternating optimization process.
Alternatively, maximization is performed.}

\item{\code{iterations}}{(\code{integer(1)})\cr
The maximum number of iterations through the parameter partition before
the alternating optimization process is terminated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Procedure-new}{\code{Procedure$new()}}
\item \href{#method-Procedure-print}{\code{Procedure$print()}}
\item \href{#method-Procedure-info}{\code{Procedure$info()}}
\item \href{#method-Procedure-next_iteration}{\code{Procedure$next_iteration()}}
\item \href{#method-Procedure-next_block}{\code{Procedure$next_block()}}
\item \href{#method-Procedure-initialize_details}{\code{Procedure$initialize_details()}}
\item \href{#method-Procedure-update_details}{\code{Procedure$update_details()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-new"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-new}{}}}
\subsection{Method \code{new()}}{
Creates a new object of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$new(verbose = TRUE, minimize = TRUE, iterations = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(\code{logical(1)})\cr
Print tracing details during the alternating optimization process?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-print"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-print}{}}}
\subsection{Method \code{print()}}{
Print details about this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-info"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$info(message, verbose = self$verbose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(\code{logical(1)})\cr
Print tracing details during the alternating optimization process?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-next_iteration"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-next_iteration}{}}}
\subsection{Method \code{next_iteration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$next_iteration(verbose = self$verbose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(\code{logical(1)})\cr
Print tracing details during the alternating optimization process?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-next_block"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-next_block}{}}}
\subsection{Method \code{next_block()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$next_block(block, verbose = self$verbose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(\code{logical(1)})\cr
Print tracing details during the alternating optimization process?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-initialize_details"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-initialize_details}{}}}
\subsection{Method \code{initialize_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$initialize_details(initial, value_at_initial, npar)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Procedure-update_details"></a>}}
\if{latex}{\out{\hypertarget{method-Procedure-update_details}{}}}
\subsection{Method \code{update_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Procedure$update_details(value, parameter, block)}\if{html}{\out{</div>}}
}

}
}
