% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_f.R
\name{set_f}
\alias{set_f}
\title{Specify function.}
\usage{
set_f(f, npar)
}
\arguments{
\item{f}{A function of \eqn{n} variables to be optimized.}

\item{npar}{The number \eqn{n} of variables of \code{f}.}
}
\value{
An object of class \code{ao_f}.
}
\description{
This function specifies the function that should be optimized.
}
\examples{
set_f(f = function(x) -x^2, npar = 1)
}
