% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_f.R
\name{set_f}
\alias{set_f}
\title{Specify optimization problem}
\usage{
set_f(f, ..., npar, lower = -Inf, upper = Inf, check = FALSE, optimizer = NULL)
}
\arguments{
\item{f}{A function to be optimized, returning a single numeric value.
Its first argument should be a numeric vector of length \code{npar}.
Additional arguments can be specified via the \code{...} argument.
Gradient or Hessian of \code{f} can be specified via attributes
\code{gradient} and \code{hessian} for the function value.
They are used for optimization if specified.}

\item{...}{Additional arguments to be passed to \code{f}.}

\item{npar}{The number of variables of \code{f}.}

\item{lower}{Lower bounds on the variables, which can be a single numeric value
(a joint bound for all parameters) or a numeric vector of length \code{npar}
(for individual bounds).}

\item{upper}{Upper bounds on the variables, analogue to \code{lower}.}

\item{check}{If \code{TRUE} checks the configuration.
This generally takes at most 20 seconds.
Set to \code{FALSE} if you are confident about the configuration to save
computation time.}

\item{optimizer}{The default optimizer is \code{\link[stats]{optim}}.
Optionally, specify an object \code{optimizer} (which can be generated by
\code{\link[ino]{set_optimizer}}) for a custom optimizer. The arguments
\code{lower} and \code{upper} are not passed to a custom optimizer.}
}
\value{
An object of class \code{ao_f}.
}
\description{
This function specifies the function to be optimized.
}
\examples{
### function to be optimized
himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2

### standard optimizer
set_f(f = himmelblau, npar = 2, lower = c(-5,-5), upper = c(5,5), check = TRUE)

### custom optimizer
custom_opt <- ino::set_optimizer(
  opt = pracma::nelder_mead, f = "fn",
  p = "x0", v = "fmin", z = "xmin", tol = 1e-6,
  crit = c("xmin", "fcount")
)
set_f(f = himmelblau, npar = 2, optimizer = custom_opt)
}
