% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optimizer.R
\name{set_optimizer}
\alias{set_optimizer}
\title{Specify optimizer.}
\usage{
set_optimizer(
  f,
  optimizer = stats::nlm,
  f_arg = "f",
  p_arg = "p",
  ...,
  check = TRUE
)
}
\arguments{
\item{f}{An object of class \code{ao_f}, i.e. the output of \code{\link{set_f}}.}

\item{optimizer}{A function which can optimize \code{f}. The default is \code{\link[stats]{nlm}}.}

\item{f_arg}{The name of the argument in \code{optimizer} that receives the function.
The default is \code{"f"}.}

\item{p_arg}{The name of the argument in \code{optimizer} that receives the initial
parameter vector. The default is \code{"p"}.}

\item{...}{Further arguments to be passed to \code{optimizer}.}

\item{check}{If \code{TRUE} checks the configuration.
This will take at most 10 seconds in most cases.
Set to \code{FALSE} if you are confident about the configuration to save
computation time.}
}
\value{
An object of class \code{ao_optimizer}.
}
\description{
This function specifies the optimization routine.
}
\examples{
set_optimizer(f = set_f(f = function(x) x^2, npar = 1), optimizer = stats::optim,
              f_arg = "fn", p_arg = "par", lower = -2, upper = -1)
}
